Incluyendo el saludo en la parte superior

+++ vistas/paginas/administradores.php
@@ -69,7 +69,10 @@ $resultado = $mysqli->query($sql);
                                             </td>
                                             <td>
                                                 <button class='btn btn-primary btn-sm'><a href="modificarAdmin.php?DNI=<?php echo $row['DNI']; ?>"><i class="far fa-edit text-white"></i></a></button>
-                                                <button class='btn btn-danger btn-sm'><a href="#" data-href="eliminarAdmin.php?DNI=<?php echo $row['DNI']; ?>" data-bs-toggle="modal" data-bs-target="#confirm-delete"><i class="fas fa-trash-alt text-white"></i></a></button>
+
+                                                <?php if ($row["DNI"] != '75182627') : ?>
+                                                    <button class='btn btn-danger btn-sm'><a href="#" data-href="eliminarAdmin.php?DNI=<?php echo $row['DNI']; ?>" data-bs-toggle="modal" data-bs-target="#confirm-delete"><i class="fas fa-trash-alt text-white"></i></a></button>
+                                                <?php endif ?>
                                             </td>
 
                                         </tr>
